feat: use Account Banner only on Home Page
