+++ src/app/legacy/containers/Navigation/index.jsx
@@ -1,17 +1,12 @@
 import { use } from 'react';
-import useToggle from '#app/hooks/useToggle';
 import { NavigationUl, NavigationLi } from '#psammead/psammead-navigation/src';
 import {
   DropdownUl,
   DropdownLi,
 } from '#psammead/psammead-navigation/src/DropdownNavigation';
-import useHydrationDetection from '#app/hooks/useHydrationDetection';
 import useClickTrackerHandler from '#app/hooks/useClickTrackerHandler';
 import useViewTracker from '#app/hooks/useViewTracker';
 import { RequestContext } from '#contexts/RequestContext';
-import { AccountContext } from '#contexts/AccountContext';
-import AccountPromotionalBanner from '#app/components/Account/AccountPromotionalBanner';
-import { ARTICLE_PAGE, HOME_PAGE } from '#app/routes/utils/pageTypes';
 import LanguageNavigation from './LanguageNavigation/lazy';
 import { ServiceContext } from '../../../contexts/ServiceContext';
 import Canonical from './index.canonical';
@@ -51,18 +46,7 @@ const renderListItems = (
   }, []);
 
 const NavigationContainer = ({ navItems, propsForTopBarOJComponent }) => {
-  const { isAmp, isLite, pageType } = use(RequestContext);
-
-  const { enabled: accountEnabled } = useToggle('account');
-  const { isSignedIn, isIdctaAvailable } = use(AccountContext);
-  const isHydrated = useHydrationDetection();
-
-  const showAccountPromoBanner =
-    isHydrated &&
-    accountEnabled &&
-    !isSignedIn &&
-    isIdctaAvailable &&
-    [ARTICLE_PAGE, HOME_PAGE].includes(pageType);
+  const { isAmp, isLite } = use(RequestContext);
 
   const { blocks = [] } = propsForTopBarOJComponent || {};
   const {
@@ -153,9 +137,7 @@ const NavigationContainer = ({ navItems, propsForTopBarOJComponent }) => {
       menuAnnouncedText={navMenuText}
       dir={dir}
       blocks={blocks}
-    >
-      {showAccountPromoBanner && <AccountPromotionalBanner />}
-    </Navigation>
+    />
   );
 };
